+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.5.0] - 2025-06-26
+### Added
+- Gemini protocol
+- More admin options
+
 ## [0.4.1] - 2025-06-25
 ### Fixed
 - Prettier data import